# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESQGRID V3",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/auth/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sos": {
      "post": {
        "summary": "Sos",
        "operationId": "sos_api_v1_sos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SOS"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/incidents": {
      "get": {
        "summary": "Incidents",
        "operationId": "incidents_api_v1_incidents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/incidents/{identifier}": {
      "get": {
        "summary": "Incident",
        "operationId": "incident_api_v1_incidents__identifier__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Transition",
        "operationId": "transition_api_v1_incidents__identifier__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{identifier}/assign": {
      "post": {
        "summary": "Assign",
        "operationId": "assign_api_v1_incidents__identifier__assign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Assignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{identifier}/messages": {
      "post": {
        "summary": "Message",
        "operationId": "message_api_v1_incidents__identifier__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Messages",
        "operationId": "messages_api_v1_incidents__identifier__messages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources": {
      "get": {
        "summary": "Resources",
        "operationId": "resources_api_v1_resources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "summary": "Alerts",
        "operationId": "alerts_api_v1_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Draft",
        "operationId": "draft_api_v1_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Alert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/{identifier}/publish": {
      "post": {
        "summary": "Publish",
        "operationId": "publish_api_v1_alerts__identifier__publish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publication"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/validate": {
      "post": {
        "summary": "Validate",
        "operationId": "validate_api_v1_telemetry_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Telemetry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/alerts/{identifier}/retract": {
      "post": {
        "summary": "Retract",
        "operationId": "retract_api_v1_alerts__identifier__retract_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Retraction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "summary": "Models",
        "operationId": "models_api_v1_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Alert": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "hazard": {
            "$ref": "#/components/schemas/Hazard"
          },
          "severity": {
            "type": "string",
            "enum": [
              "MODERATE",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "instructions": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Instructions"
          },
          "evidence_id": {
            "type": "string",
            "title": "Evidence Id"
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "hardware",
              "simulated",
              "mixed"
            ],
            "title": "Source Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Published",
              "Retracted"
            ],
            "title": "Status",
            "default": "Draft"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "zone_id",
          "hazard",
          "severity",
          "title",
          "instructions",
          "evidence_id",
          "source_mode",
          "expires_at"
        ],
        "title": "Alert"
      },
      "Assignment": {
        "properties": {
          "team_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Team Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "team_id",
          "expected_version"
        ],
        "title": "Assignment"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Location": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "source": {
            "type": "string",
            "enum": [
              "gps",
              "manual",
              "provisioned",
              "simulated"
            ],
            "title": "Source"
          },
          "accuracy_m": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy M"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "source",
          "captured_at"
        ],
        "title": "Location"
      },
...
      "Message": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Message"
      },
      "Publication": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "Publication"
      },
      "Retraction": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "Retraction"
      },
      "SOS": {
        "properties": {
          "submission_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 8,
            "title": "Submission Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "Medical",
              "Fire",
              "Flood",
              "Trapped",
              "Missing person",
              "Evacuation",
              "Infrastructure danger",
              "Other"
            ],
            "title": "Category"
          },
          "people": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "People"
          },
          "zone_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Zone Id"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "landmark": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Landmark"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "title": "Description",
            "default": ""
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "hardware",
              "simulated"
            ],
            "title": "Source Mode",
            "default": "simulated"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "submission_id",
          "category",
          "people",
          "zone_id",
          "location",
          "landmark"
        ],
        "title": "SOS"
      },
...
      "Telemetry": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "message_id": {
            "type": "string",
            "maxLength": 160,
            "minLength": 8,
            "title": "Message Id"
          },
          "device_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Device Id"
          },
          "node_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Node Id"
          },
          "zone_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Zone Id"
          },
          "boot_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Boot Id"
          },
          "sequence": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Sequence"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          },
          "firmware_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Firmware Version"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/Measurement"
            },
            "type": "array",
            "maxItems": 16,
            "minItems": 1,
            "title": "Measurements"
          },
          "battery_percent": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Percent"
          },
          "source_mode": {
            "type": "string",
            "enum": [
              "hardware",
              "simulated",
              "replay"
            ],
            "title": "Source Mode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message_id",
          "device_id",
          "node_id",
          "zone_id",
          "boot_id",
          "sequence",
          "observed_at",
          "firmware_version",
          "location",
          "measurements",
          "source_mode"
        ],
        "title": "Telemetry"
      },
      "Transition": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "maxLength": 300,
            "title": "Reason",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "expected_version"
        ],
        "title": "Transition"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}